+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* LogRecord takes tuple as `args`, provide an empty one and not a `None` (#28).
+
 # 0.8.2
 
 * Allow pyo3 0.19.